Rename component folders with capital first letter
