+++ src/components/App/App.css
@@ -9,6 +9,6 @@
   margin: auto;
 }
 
-/*#app content {*/
+/*#App Content {*/
 /*  margin-top: 8px;*/
 /*}*/
