+++ src/components/Favorites/Favorites.css
